+++ es/components/browse/SearchView.js
@@ -4,7 +4,7 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 exports.StickyFooter = StickyFooter;
-exports.SearchView = exports.ColumnCombiner = exports.SearchControllersContainer = void 0;
+exports.SearchView = exports.SearchControllersContainer = void 0;
 
 var _react = _interopRequireWildcard(require("react"));
 
@@ -60,13 +60,13 @@ function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj;
 
 function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function (obj) { return typeof obj; }; } else { _typeof = function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
 
-function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
-
-function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
-
 function _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }
 
 function _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }
+
+function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
+
+function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
 
 function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }
 
@@ -134,7 +134,7 @@ function (_React$PureComponent) {
           columnExtensionMap = _this$props.columnExtensionMap,
           currentAction = _this$props.currentAction;
 
-      var controllersAndView = _react["default"].createElement(ColumnCombiner, {
+      var controllersAndView = _react["default"].createElement(_tableCommons.ColumnCombiner, {
         context: context,
         columns: columns,
         columnExtensionMap: columnExtensionMap
@@ -144,7 +144,7 @@ function (_React$PureComponent) {
       })))));
 
       if ((0, _misc.isSelectAction)(currentAction)) {
-        controllersAndView = _react["default"].createElement(_SelectedItemsController.SelectedItemsController, null, commonControllersAndView);
+        controllersAndView = _react["default"].createElement(_SelectedItemsController.SelectedItemsController, null, controllersAndView);
       }
 
       return controllersAndView;
@@ -153,14 +153,6 @@ function (_React$PureComponent) {
 
   return SearchControllersContainer;
 }(_react["default"].PureComponent);
-/**
- * Combines `props.columns` || `props.context.columns` with `props.columnExtensionMap` to generate
- * final array of `columnDefinitions`.
- *
- * @param {{ columns: Object.<string, Object>?, context: { columns: Object.<string, Object> }, columnExtensionMap: Object.<string, Object> }} props - Props with column info.
- * @returns {JSX.Element} Clone of children, passing in `columnDefinitions` {{ field: string, ... }[]} and `defaultHiddenColumns` {Object<string, bool>}.
- */
-
 
 exports.SearchControllersContainer = SearchControllersContainer;
 
@@ -169,63 +161,45 @@ _defineProperty(SearchControllersContainer, "defaultProps", {
   'columns': null
 });
 
-var ColumnCombiner = _react["default"].memo(function (props) {
-  var children = props.children,
-      _props$columns = props.columns,
-      overridePropColumns = _props$columns === void 0 ? null : _props$columns,
-      columnExtensionMap = props.columnExtensionMap,
-      passProps = _objectWithoutProperties(props, ["children", "columns", "columnExtensionMap"]);
-
-  var contextColumns = passProps.context.columns;
-  var columns = overridePropColumns || contextColumns || null;
-
-  if (!columns) {
-    throw new Error("No columns available in context nor props. Please provide columns.");
-  }
-
-  var _useMemo = (0, _react.useMemo)(function () {
-    var columnDefinitions = (0, _tableCommons.columnsToColumnDefinitions)(columns, columnExtensionMap); // TODO: Consider changing `defaultHiddenColumnMapFromColumns` to accept array (columnDefinitions) instd of Object (columns).
-    // We currently don't put "default_hidden" property in columnExtensionMap, but could, in which case this change would be needed.
-
-    var defaultHiddenColumns = (0, _tableCommons.defaultHiddenColumnMapFromColumns)(columns);
-    return {
-      columnDefinitions: columnDefinitions,
-      defaultHiddenColumns: defaultHiddenColumns
-    };
-  }, [columns, columnExtensionMap]),
-      columnDefinitions = _useMemo.columnDefinitions,
-      defaultHiddenColumns = _useMemo.defaultHiddenColumns;
-
-  _patchedConsole.patchedConsoleInstance.log("ABV", columnDefinitions);
-
-  var propsToPass = _objectSpread({}, passProps, {
-    columnDefinitions: columnDefinitions,
-    defaultHiddenColumns: defaultHiddenColumns
-  });
-
-  return _react["default"].Children.map(children, function (child) {
-    return _react["default"].cloneElement(child, propsToPass);
-  });
-});
-
-exports.ColumnCombiner = ColumnCombiner;
-ColumnCombiner.defaultProps = {
-  "columns": null,
-  "columnExtensionMap": _tableCommons.basicColumnExtensionMap
-};
-
 var ControlsAndResults =
 /*#__PURE__*/
 function (_React$PureComponent2) {
   _inherits(ControlsAndResults, _React$PureComponent2);
 
-  /**
-   * Parses out the specific item type from `props.href` and finds the abstract item type, if any.
-   *
-   * @param {Object} props Component props.
-   * @returns {{ specificType: string, abstractType: string }} The leaf specific Item type and parent abstract type (before 'Item' in `@type` array) as strings in an object.
-   * Ex: `{ abstractType: null, specificType: "Item" }`, `{ abstractType: "Experiment", specificType: "ExperimentHiC" }`
-   */
+  _createClass(ControlsAndResults, null, [{
+    key: "searchItemTypesFromHref",
+
+    /**
+     * Parses out the specific item type from `props.href` and finds the abstract item type, if any.
+     *
+     * @param {Object} props Component props.
+     * @returns {{ specificType: string, abstractType: string }} The leaf specific Item type and parent abstract type (before 'Item' in `@type` array) as strings in an object.
+     * Ex: `{ abstractType: null, specificType: "Item" }`, `{ abstractType: "Experiment", specificType: "ExperimentHiC" }`
+     */
+    value: function searchItemTypesFromHref(href, schemas) {
+      var specificType = 'Item'; // Default
+
+      var abstractType = null; // Will be equal to specificType if no parent type.
+      // May or may not be props.href passed down from Redux (e.g. not if is EmbeddedSearchView)
+
+      var urlParts = _url["default"].parse(href, true); // Non-zero chance of having array here - though shouldn't occur unless URL entered into browser manually
+      // If we do get multiple Item types defined, we treat as if searching `type=Item` (== show `type` facet + column).
+
+
+      if (typeof urlParts.query.type === 'string') {
+        if (urlParts.query.type !== 'Item') {
+          specificType = urlParts.query.type;
+        }
+      }
+
+      abstractType = (0, _schemaTransforms.getAbstractTypeForType)(specificType, schemas) || null;
+      return {
+        specificType: specificType,
+        abstractType: abstractType
+      };
+    }
+  }]);
+
   function ControlsAndResults(props) {
     var _this2;
 
@@ -442,29 +416,6 @@ function (_React$PureComponent2) {
 
   return ControlsAndResults;
 }(_react["default"].PureComponent);
-
-_defineProperty(ControlsAndResults, "searchItemTypesFromHref", (0, _memoizeOne["default"])(function (href, schemas) {
-  var specificType = 'Item'; // Default
-
-  var abstractType = null; // Will be equal to specificType if no parent type.
-  // May or may not be props.href passed down from Redux (e.g. not if is EmbeddedSearchView)
-
-  var urlParts = _url["default"].parse(href, true); // Non-zero chance of having array here - though shouldn't occur unless URL entered into browser manually
-  // If we do get multiple Item types defined, we treat as if searching `type=Item` (== show `type` facet + column).
-
-
-  if (typeof urlParts.query.type === 'string') {
-    if (urlParts.query.type !== 'Item') {
-      specificType = urlParts.query.type;
-    }
-  }
-
-  abstractType = (0, _schemaTransforms.getAbstractTypeForType)(specificType, schemas) || null;
-  return {
-    specificType: specificType,
-    abstractType: abstractType
-  };
-}));
 
 var SearchView =
 /*#__PURE__*/
