+++ es/components/browse/components/table-commons.js
@@ -7,9 +7,9 @@ exports.sanitizeOutputValue = sanitizeOutputValue;
 exports.haveContextColumnsChanged = haveContextColumnsChanged;
 exports.defaultHiddenColumnMapFromColumns = defaultHiddenColumnMapFromColumns;
 exports.getColumnWidthFromDefinition = getColumnWidthFromDefinition;
-exports.HeadersRow = exports.ColumnSorterIcon = exports.ResultRowColumnBlockValue = exports.columnDefinitionsToScaledColumnDefinitions = exports.columnsToColumnDefinitions = exports.TableRowToggleOpenButton = exports.basicColumnExtensionMap = exports.DEFAULT_WIDTH_MAP = void 0;
+exports.HeadersRow = exports.ColumnSorterIcon = exports.ResultRowColumnBlockValue = exports.ColumnCombiner = exports.columnDefinitionsToScaledColumnDefinitions = exports.columnsToColumnDefinitions = exports.TableRowToggleOpenButton = exports.basicColumnExtensionMap = exports.DEFAULT_WIDTH_MAP = void 0;
 
-var _react = _interopRequireDefault(require("react"));
+var _react = _interopRequireWildcard(require("react"));
 
 var _propTypes = _interopRequireDefault(require("prop-types"));
 
@@ -41,6 +41,10 @@ var _typedefs = require("./../../util/typedefs");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
+function _getRequireWildcardCache() { if (typeof WeakMap !== "function") return null; var cache = new WeakMap(); _getRequireWildcardCache = function _getRequireWildcardCache() { return cache; }; return cache; }
+
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } var cache = _getRequireWildcardCache(); if (cache && cache.has(obj)) { return cache.get(obj); } var newObj = {}; if (obj != null) { var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = hasPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : null; if (desc && (desc.get || desc.set)) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj["default"] = obj; if (cache) { cache.set(obj, newObj); } return newObj; }
+
 function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -59,7 +63,15 @@ function _inherits(subClass, superClass) { if (typeof superClass !== "function" 
 
 function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function (o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
 
+function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
+
+function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
+
 function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+
+function _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }
+
+function _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }
 
 function _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _nonIterableRest(); }
 
@@ -397,6 +409,62 @@ var columnDefinitionsToScaledColumnDefinitions = (0, _memoizeOne["default"])(fun
   });
 });
 /**
+ * Combines `props.columns` || `props.context.columns` with `props.columnExtensionMap` to generate
+ * final array of `columnDefinitions`.
+ *
+ * @param {{ columns: Object.<string, Object>?, context: { columns: Object.<string, Object> }, columnExtensionMap: Object.<string, Object> }} props - Props with column info.
+ * @returns {JSX.Element} Clone of children, passing in `columnDefinitions` {{ field: string, ... }[]} and `defaultHiddenColumns` {Object<string, bool>}.
+ */
+
+exports.columnDefinitionsToScaledColumnDefinitions = columnDefinitionsToScaledColumnDefinitions;
+
+var ColumnCombiner = _react["default"].memo(function (props) {
+  var children = props.children,
+      _props$columns = props.columns,
+      overridePropColumns = _props$columns === void 0 ? null : _props$columns,
+      columnExtensionMap = props.columnExtensionMap,
+      passProps = _objectWithoutProperties(props, ["children", "columns", "columnExtensionMap"]);
+
+  var contextColumns = passProps.context.columns;
+  var columns = overridePropColumns || contextColumns || null;
+
+  if (!columns) {
+    throw new Error("No columns available in context nor props. Please provide columns.");
+  }
+
+  var _useMemo = (0, _react.useMemo)(function () {
+    var columnDefinitions = columnsToColumnDefinitions(columns, columnExtensionMap); // TODO: Consider changing `defaultHiddenColumnMapFromColumns` to accept array (columnDefinitions) instd of Object (columns).
+    // We currently don't put "default_hidden" property in columnExtensionMap, but could, in which case this change would be needed.
+
+    var defaultHiddenColumns = defaultHiddenColumnMapFromColumns(columns);
+    return {
+      columnDefinitions: columnDefinitions,
+      defaultHiddenColumns: defaultHiddenColumns
+    };
+  }, [columns, columnExtensionMap]),
+      columnDefinitions = _useMemo.columnDefinitions,
+      defaultHiddenColumns = _useMemo.defaultHiddenColumns;
+
+  console.log("ABV", columnDefinitions);
+
+  var propsToPass = _objectSpread({}, passProps, {
+    columnDefinitions: columnDefinitions,
+    defaultHiddenColumns: defaultHiddenColumns
+  });
+
+  return _react["default"].Children.map(children, function (child) {
+    return _react["default"].cloneElement(child, propsToPass);
+  });
+});
+
+exports.ColumnCombiner = ColumnCombiner;
+ColumnCombiner.defaultProps = {
+  "columns": null,
+  // Passed in as prop or defaults to context.columns
+  "columnExtensionMap": basicColumnExtensionMap,
+  "context": {}
+};
+/**
  * Determine the typical column width, given current browser width. Defaults to large width if server-side.
  *
  * @param {ColumnDefinition} columnDefinition - JSON of column definition, should have widthMap or width or baseWidth.
@@ -405,8 +473,6 @@ var columnDefinitionsToScaledColumnDefinitions = (0, _memoizeOne["default"])(fun
  * @param {number} [windowWidth=null] - Current window width.
  * @returns {string|number} Width for div column block to be used at current screen/browser size.
  */
-
-exports.columnDefinitionsToScaledColumnDefinitions = columnDefinitionsToScaledColumnDefinitions;
 
 function getColumnWidthFromDefinition(columnDefinition) {
   var mounted = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : true;
